backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0

# LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LLM + Embeddings (Gemini)
langchain-google-genai>=2.0.0

# Vector store
langchain-chroma>=0.2.0
chromadb>=0.5.0

# HTTP client
httpx>=0.27.0

# Configuration
pydantic-settings>=2.5.0
python-dotenv>=1.0.0

# Utilities
tenacity>=9.0.0
tqdm>=4.66.0
